refactor(types): types are separated
